collector/requirements.txt
requests==2.32.5
python-frontmatter==1.1.0
